Added product search filtering
