+++ online-shopping-website/src/Components/SearchBar.js
@@ -1,0 +1,20 @@
+import { TextField } from "@mui/material";
+
+export const SearchBar = (props) => {
+    let filters = props.filters;
+    const filterProducts = props.filterProducts;
+
+    const handleSearchChange = (event) => {
+        filters.searchQuery = event.target.value;
+        filterProducts();
+    }
+
+    return (
+        <TextField 
+            label="Search Products" 
+            variant="outlined"
+            type="search"
+            onChange={(e) => handleSearchChange(e)}
+        />
+    );
+}
